Simulation Writer (#9)
* Add bound values for geenrating particles position's

* Add checking for infinite looping on initalizing particles

* Add method simulate()

* Remove method simulate()

* Add class SimulationWriter

* Add closing for DataOuputStream object

* Add throwing exception in case of infinite loop

* Add block try-catch for infinite loop exception

* Add block try-catch for initialDistribution()

* Add specifier 'throws' to method initialDistribution()

* Replace java.io.* with required parts of java.io

* Add specifier 'final' to SimulationWriter constructor arguments

* Remove try-catch block in initialDistribuition() for unused

* Changed try-catch block with codestyle

* Change writing's order of arguments and replace writeDouble() with writeFloat()

* Change return value type of getInitialSpeed() from double to float

* Add throwing exception in case if box so small for the particles

* Add particle's initalazing at step of overlapsing's checking

* Change loop variables for correct iterations

* Remove try-catch block for catching initialDistribution() exceptions

* Remove vartiable currTime for unused

* Add method calculateTimeStep()

* Add specifier static to method calculateTimeStep()

* Change method's calculateTimeStep() return value from double to long

* Changes number of the states which writes to file

* Change delta time for states

* Correct comparator for particles in nextTimeStep()

* Change precision method's calculateTimeStep return value

* Add whiteship after typecasts

* Add tests for method initialDistribution()

* Add test for method nextTimeStep()

* Change bounds for X-coordinate around hole

* Add sorting of array before comapring

* Change cup of infinite loop counter from 10000000 to 20000000

* Correct particles indices when filling right box part

* Add newline to the end of SimulatorTest.java

* Move counter to do-while scope

+++ src/main/java/eu/mivrenik/particles/model/ExperimentSettings.java
@@ -110,7 +110,7 @@ public final class ExperimentSettings {
         return particleCountLeft + particleCountRight;
     }
 
-    public double getInitialSpeed() {
+    public float getInitialSpeed() {
         return initialSpeed;
     }
